chore: moved func to utils
